# contracts/openapi/public-api.yaml
openapi: 3.1.0
...
info:
  title: EduTwin Public API
  version: 2.0.0
  license:
    name: Course demonstration use
    identifier: LicenseRef-Course-Demonstration
  description: |
    EduTwin browser-facing contract. MySQL is the system of record. Redis values,
    task state, and server-sent events are recoverable projections. Every traced
    response identifies the accepted answer event, analysis job, preallocated or
    persisted snapshot, data versions, and requested and effective model versions.
    Authentication uses a stateless JWT with one active role and all granted roles.
    Role, password, and account changes advance the token version and invalidate older
    tokens. Logout consists of discarding the JWT in the browser.
servers:
  - url: http://127.0.0.1:8080
    description: Local Compose deployment
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/student/courses/{courseId}/analytics:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - $ref: '#/components/parameters/AnalyticsPeriod'
    get:
      tags: [Student Twin]
      operationId: getStudentAnalytics
      summary: Return self-service learning trends and behavior analytics
      x-rbac: [STUDENT_SELF]
      responses:
        '200':
          description: Student learning analytics for the selected period.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAnalytics'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/courses/{courseId}/students/{studentId}/twin/current:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - $ref: '#/components/parameters/StudentId'
    get:
      tags: [Student Twin]
      operationId: getCurrentTwin
      summary: Return the current immutable twin snapshot
      description: Students may read only their own twin; teachers may read enrolled students in assigned courses.
      x-rbac: [STUDENT_SELF, TEACHER_ASSIGNED_TO_COURSE]
      responses:
        '200':
          description: Current twin state, read from MySQL or a validated Redis projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/courses/{courseId}/students/{studentId}/twin/history:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - $ref: '#/components/parameters/StudentId'
      - $ref: '#/components/parameters/BeforeVersion'
      - $ref: '#/components/parameters/Limit'
    get:
      tags: [Student Twin]
      operationId: getTwinHistory
      summary: Return immutable twin snapshot history
      description: Results are ordered by snapshotVersion descending.
      x-rbac: [STUDENT_SELF, TEACHER_ASSIGNED_TO_COURSE]
      responses:
        '200':
          description: Snapshot history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/risk/cases:
    get:
      tags: [Risk Interventions]
      operationId: listRiskCases
      summary: List risk cases within the active teacher or counselor scope
      x-rbac: [TEACHER, COUNSELOR]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/RiskCaseStatus' } }
        - { name: assignedTo, in: query, schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200': { description: Paged risk cases., content: { application/json: { schema: { $ref: '#/components/schemas/RiskCasePage' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Risk Interventions]
      operationId: createRiskCase
      summary: Create a manual risk case in the active role scope
      x-rbac: [TEACHER, COUNSELOR]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskCaseCreateRequest' } } }
      responses:
        '200': { description: Risk case created., content: { application/json: { schema: { $ref: '#/components/schemas/RiskCaseDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/risk/cases/{caseId}:
    parameters:
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Risk Interventions]
      operationId: getRiskCase
      summary: Return a risk case with assignment, note, and action history
      x-rbac: [TEACHER, COUNSELOR]
      responses:
        '200': { description: Risk case detail., content: { application/json: { schema: { $ref: '#/components/schemas/RiskCaseDetail' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/risk/cases/{caseId}/assignment:
    parameters:
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Risk Interventions]
      operationId: assignRiskCase
      summary: Optimistically assign a risk case to an eligible user
      x-rbac: [TEACHER, COUNSELOR]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskAssignmentRequest' } } }
      responses:
        '200': { description: Risk case assigned., content: { application/json: { schema: { $ref: '#/components/schemas/RiskCaseDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/risk/cases/{caseId}/status:
    parameters:
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Risk Interventions]
      operationId: transitionRiskCase
      summary: Optimistically transition a risk case lifecycle state
      x-rbac: [TEACHER, COUNSELOR]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskStatusRequest' } } }
      responses:
        '200': { description: Risk case transitioned., content: { application/json: { schema: { $ref: '#/components/schemas/RiskCaseDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/risk/cases/{caseId}/notes:
    parameters:
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Risk Interventions]
      operationId: addRiskCaseNote
      summary: Add an internal note to a risk case
      x-rbac: [TEACHER, COUNSELOR]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskNoteRequest' } } }
      responses:
        '200': { description: Risk case with note., content: { application/json: { schema: { $ref: '#/components/schemas/RiskCaseDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/risk/cases/{caseId}/actions:
    parameters:
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Risk Interventions]
      operationId: addRiskCaseAction
      summary: Add a student-visible action item to a risk case
      x-rbac: [TEACHER, COUNSELOR]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskActionRequest' } } }
      responses:
        '200': { description: Risk case with action item., content: { application/json: { schema: { $ref: '#/components/schemas/RiskCaseDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/risk/cases/{caseId}/actions/{actionId}/status:
    parameters:
      - { name: caseId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: actionId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Risk Interventions]
      operationId: updateRiskCaseActionStatus
      summary: Transition a risk action item and record its result
      x-rbac: [TEACHER, COUNSELOR]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskActionStatusRequest' } } }
      responses:
        '200': { description: Risk case with updated action., content: { application/json: { schema: { $ref: '#/components/schemas/RiskCaseDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/risk/me/actions:
    get:
      tags: [Risk Interventions]
      operationId: listMyRiskActions
      summary: List non-cancelled action items assigned to the current student
      x-rbac: [STUDENT]
      responses:
        '200':
          description: Student action items.
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/StudentRiskAction' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/risk/me/actions/{actionId}/feedback:
    parameters:
      - { name: actionId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Risk Interventions]
      operationId: addMyRiskActionFeedback
      summary: Add feedback to an owned active action item
      x-rbac: [STUDENT]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskFeedbackRequest' } } }
      responses:
        '200': { description: Student action with feedback., content: { application/json: { schema: { $ref: '#/components/schemas/StudentRiskAction' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateless JWT containing one role plus userId and authorization-version claims; logout discards the token client-side.
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    StudentId:
      name: studentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BeforeVersion:
      name: beforeVersion
      in: query
      required: false
      description: Exclusive upper snapshot version bound.
      schema:
        type: integer
        format: int64
        minimum: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    AnalyticsPeriod:
      name: period
      in: query
      required: false
      schema:
        type: string
        enum: [7D, 30D, TERM]
        default: 30D
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Authenticated principal lacks the required role or scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found within the authorized scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Idempotency conflict, stale snapshot, or immutable-version conflict.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    DataVersionRef:
      type: object
      additionalProperties: false
      required: [sourceId, version, manifestSha256]
      properties:
        sourceId:
          type: string
          enum: [ASSISTMENTS_2009_2010_SKILL_BUILDER_CORRECTED, OULAD, EDUTWIN_DEMO]
        version:
          type: string
          minLength: 1
        manifestSha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
    ModelVersionRef:
      type: object
      additionalProperties: false
      required: [purpose, family, modelName, modelVersion, artifactSha256, calibratorVersion]
      properties:
        purpose:
          type: string
          enum: [MASTERY, NEXT_CORRECT, RISK, EXPLANATION, DIAGNOSIS, PLAN_RULES]
        family:
          type: string
          enum: [IRT, BKT, DKT, AKT, LOGISTIC_REGRESSION, LIGHTGBM, CATBOOST, SHAP, DEEPSEEK, OPENAI_COMPATIBLE, RULE, TEMPLATE]
        modelName:
          type: string
          minLength: 1
        modelVersion:
          type: string
          minLength: 1
        artifactSha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        calibratorVersion:
          type: [string, 'null']
          minLength: 1
    TraceRef:
      type: object
      additionalProperties: false
      required:
        - correlationId
        - answerEventId
        - analysisJobId
        - snapshotId
        - dataVersions
        - requestedModelVersions
        - effectiveModelVersions
      properties:
        correlationId:
          type: string
          format: uuid
        answerEventId:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
        snapshotId:
          type: string
          format: uuid
          description: Preallocated for QUEUED jobs and persisted unchanged on completion.
        dataVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/DataVersionRef'
        requestedModelVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ModelVersionRef'
        effectiveModelVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ModelVersionRef'
    RiskPrediction:
      type: object
      additionalProperties: false
      required:
        - probability
        - riskBand
        - mediumThreshold
        - highThreshold
        - predictedAt
      properties:
        trace:
          $ref: '#/components/schemas/TraceRef'
          x-edutwin-visibility: internal
        probability:
          type: number
          minimum: 0
          maximum: 1
        calibrated:
          type: boolean
          x-edutwin-visibility: internal
        riskBand:
          $ref: '#/components/schemas/RiskBand'
        mediumThreshold:
          type: number
          minimum: 0
          maximum: 1
        highThreshold:
          type: number
          minimum: 0
          maximum: 1
        baseValue:
          type: number
          description: Selected risk model base value in log-odds space.
          x-edutwin-visibility: internal
        featureContractVersion:
          type: string
          minLength: 1
          x-edutwin-visibility: internal
        predictedAt:
          type: string
          format: date-time
        degraded:
          type: boolean
          x-edutwin-visibility: internal
    RiskBand:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    SkillMastery:
      type: object
      additionalProperties: false
      required: [skillId, skillName, probability]
      properties:
        skillId:
          type: string
          format: uuid
        skillName:
          type: string
          minLength: 1
        probability:
          type: number
          minimum: 0
          maximum: 1
        estimator:
          type: string
          enum: [BKT, RULE_FALLBACK]
          x-edutwin-visibility: internal
    TwinState:
      type: object
      additionalProperties: false
      required:
        - studentId
        - courseId
        - snapshotVersion
        - capturedAt
        - mastery
        - engagementScore
        - stabilityScore
        - nextCorrectProbability
        - risk
        - planCompletionRate
      properties:
        trace:
          $ref: '#/components/schemas/TraceRef'
          x-edutwin-visibility: internal
        studentId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        snapshotVersion:
          type: integer
          format: int64
          minimum: 1
        capturedAt:
          type: string
          format: date-time
        mastery:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SkillMastery'
        engagementScore:
          type: number
          minimum: 0
          maximum: 1
        stabilityScore:
          type: number
          minimum: 0
          maximum: 1
        nextCorrectProbability:
          type: number
          minimum: 0
          maximum: 1
        risk:
          $ref: '#/components/schemas/RiskPrediction'
        planCompletionRate:
          type: number
          minimum: 0
          maximum: 1
        degraded:
          type: boolean
          x-edutwin-visibility: internal
        degradationReasons:
          type: array
          uniqueItems: true
          x-edutwin-visibility: internal
          items:
            type: string
            enum: [KNOWLEDGE_MODEL_UNAVAILABLE, RISK_MODEL_UNAVAILABLE, MODEL_TIMEOUT, MODEL_CONTRACT_ERROR]
    TwinHistory:
      type: object
      additionalProperties: false
      required: [items, total, nextBeforeVersion]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TwinState'
        total:
          type: integer
          minimum: 0
        nextBeforeVersion:
          type: [integer, 'null']
          format: int64
          minimum: 1
...
    ActivityPoint:
      type: object
      additionalProperties: false
      required: [date, answerCount, activeStudentCount, correctRate]
      properties:
        date:
          type: string
          format: date
        answerCount:
          type: integer
          minimum: 0
        activeStudentCount:
          type: integer
          minimum: 0
        correctRate:
          type: number
          minimum: 0
          maximum: 1
    BehaviorDistributionItem:
      type: object
      additionalProperties: false
      required: [behaviorType, eventCount, studentCount, share]
      properties:
        behaviorType: { type: string }
        eventCount: { type: integer, minimum: 0 }
        studentCount: { type: integer, minimum: 0 }
        share: { type: number, minimum: 0, maximum: 1 }
    ActivityHeatmapPoint:
      type: object
      additionalProperties: false
      required: [weekday, hour, eventCount]
      properties:
        weekday: { type: integer, minimum: 1, maximum: 7 }
        hour: { type: integer, minimum: 0, maximum: 23 }
        eventCount: { type: integer, minimum: 0 }
...
    StudentAnalytics:
      type: object
      additionalProperties: false
      required: [courseId, studentId, period, generatedAt, answerCount, correctRate, activeDays, totalDurationMinutes, currentMastery, currentRisk, currentEngagement, snapshotTrend, activityTrend, behaviorDistribution, activityHeatmap]
      properties:
        courseId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        period: { type: string, enum: [7D, 30D, TERM] }
        generatedAt: { type: string, format: date-time }
        answerCount: { type: integer, minimum: 0 }
        correctRate: { type: number, minimum: 0, maximum: 1 }
        activeDays: { type: integer, minimum: 0 }
        totalDurationMinutes: { type: integer, minimum: 0 }
        currentMastery: { type: number, minimum: 0, maximum: 1 }
        currentRisk: { type: number, minimum: 0, maximum: 1 }
        currentEngagement: { type: number, minimum: 0, maximum: 1 }
        snapshotTrend: { type: array, items: { $ref: '#/components/schemas/TwinState' } }
        activityTrend: { type: array, items: { $ref: '#/components/schemas/ActivityPoint' } }
        behaviorDistribution: { type: array, items: { $ref: '#/components/schemas/BehaviorDistributionItem' } }
        activityHeatmap: { type: array, items: { $ref: '#/components/schemas/ActivityHeatmapPoint' } }
...
    RiskCaseStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
    RiskActionStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED, CANCELLED]
    RiskCaseSummary:
      type: object
      additionalProperties: false
      required: [caseId, courseId, studentId, studentName, courseTitle, triggerType, riskBand, title, summary, status, priority, assignedTo, assigneeName, version, createdAt, updatedAt, resolvedAt, closedAt]
      properties:
        caseId: { type: string, format: uuid }
        courseId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        studentName: { type: string }
        courseTitle: { type: string }
        triggerType: { type: string }
        riskBand: { $ref: '#/components/schemas/RiskBand' }
        title: { type: string }
        summary: { type: string }
        status: { $ref: '#/components/schemas/RiskCaseStatus' }
        priority: { type: string }
        assignedTo: { type: [string, 'null'], format: uuid }
        assigneeName: { type: [string, 'null'] }
        version: { type: integer, format: int64, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
        closedAt: { type: [string, 'null'], format: date-time }
    RiskCasePage:
      type: object
      additionalProperties: false
      required: [items, total, page, size]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RiskCaseSummary' } }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
    RiskAssignment:
      type: object
      additionalProperties: false
      required: [assignmentId, fromAssignee, toAssignee, changedBy, reason, createdAt]
      properties:
        assignmentId: { type: string, format: uuid }
        fromAssignee: { type: [string, 'null'], format: uuid }
        toAssignee: { type: string, format: uuid }
        changedBy: { type: string, format: uuid }
        reason: { type: string }
        createdAt: { type: string, format: date-time }
    RiskInternalNote:
      type: object
      additionalProperties: false
      required: [noteId, authorId, authorName, body, createdAt]
      properties:
        noteId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        authorName: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    RiskFeedback:
      type: object
      additionalProperties: false
      required: [feedbackId, studentId, body, createdAt]
      properties:
        feedbackId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    RiskActionItem:
      type: object
      additionalProperties: false
      required: [actionItemId, caseId, studentId, title, description, status, dueAt, resultSummary, createdAt, startedAt, completedAt, feedback]
      properties:
        actionItemId: { type: string, format: uuid }
        caseId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/RiskActionStatus' }
        dueAt: { type: [string, 'null'], format: date-time }
        resultSummary: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        feedback: { type: array, items: { $ref: '#/components/schemas/RiskFeedback' } }
    RiskCaseDetail:
      type: object
      additionalProperties: false
      required: [riskCase, assignments, internalNotes, actionItems]
      properties:
        riskCase: { $ref: '#/components/schemas/RiskCaseSummary' }
        assignments: { type: array, items: { $ref: '#/components/schemas/RiskAssignment' } }
        internalNotes: { type: array, items: { $ref: '#/components/schemas/RiskInternalNote' } }
        actionItems: { type: array, items: { $ref: '#/components/schemas/RiskActionItem' } }
    RiskCaseCreateRequest:
      type: object
      additionalProperties: false
      required: [courseId, studentId, title, summary]
      properties:
        courseId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 180 }
        summary: { type: string, minLength: 1, maxLength: 1000 }
    RiskAssignmentRequest:
      type: object
      additionalProperties: false
      required: [assigneeId, reason, expectedVersion]
      properties:
        assigneeId: { type: string, format: uuid }
        reason: { type: string, minLength: 1, maxLength: 300 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    RiskStatusRequest:
      type: object
      additionalProperties: false
      required: [targetStatus, reason, expectedVersion]
      properties:
        targetStatus: { $ref: '#/components/schemas/RiskCaseStatus' }
        reason: { type: string, minLength: 1, maxLength: 300 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    RiskNoteRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 2000 }
    RiskActionRequest:
      type: object
      additionalProperties: false
      required: [title, description, dueAt]
      properties:
        title: { type: string, minLength: 1, maxLength: 180 }
        description: { type: string, minLength: 1, maxLength: 2000 }
        dueAt: { type: string, format: date-time }
    RiskActionStatusRequest:
      type: object
      additionalProperties: false
      required: [targetStatus]
      properties:
        targetStatus: { $ref: '#/components/schemas/RiskActionStatus' }
        resultSummary: { type: [string, 'null'], maxLength: 1000 }
    RiskFeedbackRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 1000 }
    StudentRiskAction:
      type: object
      additionalProperties: false
      required: [actionItemId, caseId, courseId, courseTitle, title, description, status, dueAt, resultSummary, feedback]
      properties:
        actionItemId: { type: string, format: uuid }
        caseId: { type: string, format: uuid }
        courseId: { type: string, format: uuid }
        courseTitle: { type: string }
        title: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/RiskActionStatus' }
        dueAt: { type: [string, 'null'], format: date-time }
        resultSummary: { type: [string, 'null'] }
        feedback: { type: array, items: { $ref: '#/components/schemas/RiskFeedback' } }
...
    FieldViolation:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
    ProblemDetail:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, traceId, violations]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        traceId:
          type: string
          minLength: 1
        violations:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'